refactor(server): extract shared song field validation and playback filter building (#778)
